DRD-2547: Google Cast Icon Placement
fix issue when chromecast button is displaying when no chromecast device available.
The reason of issue is that current action bar menu implementation
suppose full redrawing on each open/close of drawer.

The fix is to add access to count of currently available devices
which helps to determine visibility of chromecast button during redrawing.