DRD-4354: The Google+ Sign up option doesnt show at the login gate.
The reason of issue is in incompatibility of installed versions of Google Play Services and version of library included to iHeartRadio.
Library in iHeartRadio wants to  newer version of Google Play Services.
At the same time Google Play doesn't provide this version.
Due to this fact google play services is not available in app.

The fix is to set play-services library version as the latest available on the Google Play.